--- frontend/src/components/pages/CampOverview/CampOverview.css
@@ -1,3 +1,0 @@
-.camper-checkbox {
-  accent-color: #468740;
-}
