+++ frontend/src/components/pages/CampOverview/CampersTable/CampersTable.tsx
@@ -17,7 +17,7 @@ import {
   Tag,
   Box,
   useDisclosure,
-  AlertProps,
+  useToast,
 } from "@chakra-ui/react";
 import { DownloadIcon, SearchIcon } from "@chakra-ui/icons";
 
@@ -31,12 +31,12 @@ import textStyles from "../../../../theme/textStyles";
 import CampersTableKebabMenu from "./CampersTableKebabMenu";
 
 import EditCamperModal from "../EditCamperModal";
-import MoveCamperModal from "../MoveCampersModal";
+import MoveCamperModal from "../MoveCamperModal";
 import ViewCamperModal from "../ViewCamperModal";
 
 import { CampSession } from "../../../../types/CampsTypes";
 import CamperAPIClient from "../../../../APIClients/CamperAPIClient";
-import Alert from "../../../common/Alert";
+import CampsAPIClient from "../../../../APIClients/CampsAPIClient";
 
 const ExportButton = (): JSX.Element => {
   return (
@@ -58,13 +58,11 @@ const ExportButton = (): JSX.Element => {
 };
 
 const CampersTable = ({
-  currentCampSession,
   campers,
   campSession,
   campSessionCapacity,
   updateCamp,
 }: {
-  currentCampSession: number;
   campers: Camper[];
   campSession: CampSession;
   campSessionCapacity: number;
@@ -73,6 +71,10 @@ const CampersTable = ({
   const [search, setSearch] = React.useState("");
   const [selectedFilter, setSelectedFilter] = React.useState<Filter>(
     Filter.ALL,
+  );
+  const [campSessions, setCampSessions] = React.useState<CampSession[]>([]);
+  const [selectedCampSession, setSelectedCampSession] = React.useState<string>(
+    campSession.id,
   );
 
   const {
@@ -93,10 +95,7 @@ const CampersTable = ({
     onClose: viewModalOnClose,
   } = useDisclosure();
 
-  const [alertInfo, setAlertInfo] = React.useState<{
-    status: AlertProps["status"];
-    description: string;
-  } | null>(null);
+  const toast = useToast();
 
   const tableData = React.useMemo(() => {
     let filteredCampers;
@@ -149,32 +148,49 @@ const CampersTable = ({
     setCamperDetailsCount(tempDetailsCount);
   }, [campers]);
 
+  React.useEffect(() => {
+    CampsAPIClient.getCampById(campSession.camp).then((camp) => {
+      setCampSessions(camp.campSessions);
+    });
+  }, [campSession.camp]);
+
   // Function to be called by the Move Campers Modal to update the campers
-  const moveCampers = (camperIds: string[], campSessionId: string) => {
-    CamperAPIClient.moveCampersToCampSession(camperIds, campSessionId).then(
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      (response: any) => {
-        if (Object.keys(response).includes("isAxiosError")) {
-          setAlertInfo({
-            status: "error",
-            description: response.response.data.error,
-          });
-        } else {
-          setAlertInfo({
-            status: "success",
-            description: `${(response as Camper[])
-              .slice(0, -1)
-              .map((camper) => `${camper.firstName} ${camper.lastName}`)
-              .join(", ")}${response.length !== 1 ? " and " : ""}${
-              response[response.length - 1].firstName
-            } ${response[response.length - 1].lastName} ${
-              response.length !== 1 ? "have" : "has"
-            } been moved to Session ${currentCampSession + 1}.`,
-          });
-          updateCamp();
-        }
-      },
+  const moveCampers = async (camperIds: string[], campSessionId: string) => {
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    const res: any = await CamperAPIClient.moveCampersToCampSession(
+      camperIds,
+      campSessionId,
     );
+
+    if (Object.keys(res).includes("isAxiosError")) {
+      toast({
+        position: "bottom",
+        title: res.response.data.error ?? res.response.data,
+        status: "error",
+        duration: 3000,
+        isClosable: true,
+      });
+    } else {
+      toast({
+        position: "bottom",
+        title: `${(res as Camper[])
+          .slice(0, -1)
+          .map((camper) => `${camper.firstName} ${camper.lastName}`)
+          .join(", ")}${res.length !== 1 ? " and " : ""}${
+          res[res.length - 1].firstName
+        } ${res[res.length - 1].lastName} ${
+          res.length !== 1 ? "have" : "has"
+        } been moved to Session ${
+          campSessions
+            .map((campSessionItem) => campSessionItem.id)
+            .indexOf(selectedCampSession) + 1
+        }.`,
+        status: "success",
+        duration: 3000,
+        isClosable: true,
+      });
+      updateCamp();
+    }
   };
 
   return (
@@ -310,6 +326,9 @@ const CampersTable = ({
               <MoveCamperModal
                 camper={selectedCamper}
                 campSession={campSession}
+                campSessions={campSessions}
+                selectedCampSession={selectedCampSession}
+                setSelectedCampSession={setSelectedCampSession}
                 moveCampers={moveCampers}
                 moveCamperModalIsOpen={moveModalIsOpen}
                 moveCamperModalOnClose={moveModalOnClose}
@@ -338,9 +357,6 @@ const CampersTable = ({
           <Text>Check back later!</Text>
         </VStack>
       )}
-      {alertInfo ? (
-        <Alert status={alertInfo.status} description={alertInfo.description} />
-      ) : null}
     </Box>
   );
 };
